Move Dummy driver to apropriate namespace
Use @inheritDoc on methods docblocks
